Added FM synth and Trigger Sample based on groove~ [beta] + soundpool accepts single files
Added
- FM synth [monophonic, beta]
- Trigger module based on polybuffer~ and groove~ with ADSR~
- simple data recorder [beta, not great]

Modified
- synth.ctrl.soundpool works with single files now too (in addition to
folder) via drag and drop

Moved
- Old wax source file moved into old folder, do not delete for now.

Gitignore
- Sublime project files
