+++ .gitignore
@@ -1,0 +1,26 @@
+
+mod-interfaces/wax/Configuration/testRecWax9/GEST01_TRIAL02_12_21_2015_17_52_7__MAG.txt
+
+mod-interfaces/wax/Configuration/testRecWax9/GEST01_TRIAL02_12_21_2015_17_52_7__GYRO.txt
+
+mod-interfaces/wax/Configuration/testRecWax9/GEST01_TRIAL02_12_21_2015_17_52_7__acc.txt
+
+mod-interfaces/wax/Configuration/testRecWax9/GEST01_TRIAL01_12_21_2015_17_51_13__MAG.txt
+
+mod-interfaces/wax/Configuration/testRecWax9/GEST01_TRIAL01_12_21_2015_17_51_13__GYRO.txt
+
+mod-interfaces/wax/Configuration/testRecWax9/GEST01_TRIAL01_12_21_2015_17_51_13__acc.txt
+
+*.sublime-project
+
+*.sublime-workspace
+
+mod-interfaces/wax/Configuration/waxrec/.DS_Store
+
+mod-analysis/.DS_Store
+
+examples/sounds/.DS_Store
+
+examples/.DS_Store
+
+mod-synthesis/.DS_Store
