--- mod-interfaces/wax/Configuration/Others/slip2udp/Makefile.copy
@@ -1,14 +1,0 @@
-# Makefile
-
-CC=gcc
-CFLAGS = -isysroot /Developer/SDKs/MacOSX10.5.sdk -mmacosx-version-min=10.5
-
-
-
-all: build
-
-build: Makefile slip2udp.c
-	$(CC) -o slip2udp slip2udp.c
-
-clean:
-	rm -f slip2udp.o core
