--- mod-interfaces/wax/Configuration/scripts/wax-RX_p8200.command
@@ -1,7 +1,0 @@
-#!/bin/bash
-clear
-cd ~/Dropbox/MachineLearningPlayground/MaxMspPatches/hardware/WAX/waxrec/
-./waxrec /dev/tty.usbmodem* -osc localhost:8200
-
-
-
