--- mod-interfaces/wax/Configuration/scripts/wax-RX_p8200.sh
@@ -1,6 +1,0 @@
-#!/bin/bash
-#cd ~/Dropbox/MachineLearningPlayground/MaxMspPatches/hardware/WAX/waxrec/
-~/Dropbox/SID-Workshops/MaxMSP/hardware/WAX/waxrec/waxrec /dev/tty.usbmodem* -osc localhost:8200
-
-
-
