+++ mod-synthesis/synth.ctrl.soundpool.maxpat
@@ -2,31 +2,54 @@
 	"patcher" : 	{
 		"fileversion" : 1,
 		"appversion" : 		{
-			"major" : 6,
+			"major" : 7,
 			"minor" : 0,
-			"revision" : 5
+			"revision" : 6,
+			"architecture" : "x86",
+			"modernui" : 1
 		}
 ,
-		"rect" : [ 60.0, 46.0, 650.0, 665.0 ],
+		"rect" : [ 60.0, 79.0, 650.0, 665.0 ],
 		"bglocked" : 0,
 		"openinpresentation" : 1,
 		"default_fontsize" : 12.0,
 		"default_fontface" : 0,
 		"default_fontname" : "Arial",
-		"gridonopen" : 0,
+		"gridonopen" : 1,
 		"gridsize" : [ 15.0, 15.0 ],
-		"gridsnaponopen" : 0,
+		"gridsnaponopen" : 1,
+		"objectsnaponopen" : 1,
 		"statusbarvisible" : 2,
 		"toolbarvisible" : 1,
+		"lefttoolbarpinned" : 0,
+		"toptoolbarpinned" : 0,
+		"righttoolbarpinned" : 0,
+		"bottomtoolbarpinned" : 0,
+		"toolbars_unpinned_last_save" : 0,
+		"tallnewobj" : 0,
 		"boxanimatetime" : 200,
-		"imprint" : 0,
 		"enablehscroll" : 1,
 		"enablevscroll" : 1,
 		"devicewidth" : 0.0,
 		"description" : "",
 		"digest" : "",
 		"tags" : "",
+		"style" : "",
+		"subpatcher_template" : "",
 		"boxes" : [ 			{
+				"box" : 				{
+					"id" : "obj-8",
+					"maxclass" : "message",
+					"numinlets" : 2,
+					"numoutlets" : 1,
+					"outlettype" : [ "" ],
+					"patching_rect" : [ 539.0, 173.0, 50.0, 22.0 ],
+					"style" : "",
+					"text" : "1"
+				}
+
+			}
+, 			{
 				"box" : 				{
 					"fontname" : "Arial",
 					"fontsize" : 12.0,
@@ -35,7 +58,8 @@
 					"numinlets" : 1,
 					"numoutlets" : 1,
 					"outlettype" : [ "" ],
-					"patching_rect" : [ 181.0, 183.0, 91.0, 20.0 ],
+					"patching_rect" : [ 181.0, 183.0, 91.0, 22.0 ],
+					"style" : "",
 					"text" : "loadmess clear"
 				}
 
@@ -47,7 +71,8 @@
 					"maxclass" : "outlet",
 					"numinlets" : 1,
 					"numoutlets" : 0,
-					"patching_rect" : [ 307.25, 283.0, 25.0, 25.0 ]
+					"patching_rect" : [ 307.25, 283.0, 25.0, 25.0 ],
+					"style" : ""
 				}
 
 			}
@@ -59,7 +84,8 @@
 					"numinlets" : 0,
 					"numoutlets" : 1,
 					"outlettype" : [ "" ],
-					"patching_rect" : [ 282.0, 20.0, 25.0, 25.0 ]
+					"patching_rect" : [ 282.0, 20.0, 25.0, 25.0 ],
+					"style" : ""
 				}
 
 			}
@@ -77,10 +103,10 @@
 					"presentation_rect" : [ 88.0, 28.5, 15.0, 15.0 ],
 					"saved_attribute_attributes" : 					{
 						"valueof" : 						{
-							"parameter_mmax" : 1.0,
-							"parameter_type" : 2,
-							"parameter_shortname" : "live.toggle",
 							"parameter_longname" : "live.toggle[2]",
+							"parameter_shortname" : "live.toggle",
+							"parameter_type" : 2,
+							"parameter_mmax" : 1.0,
 							"parameter_enum" : [ "off", "on" ]
 						}
 
@@ -104,10 +130,10 @@
 					"presentation_rect" : [ 88.0, 6.0, 15.0, 15.0 ],
 					"saved_attribute_attributes" : 					{
 						"valueof" : 						{
-							"parameter_mmax" : 1.0,
-							"parameter_type" : 2,
-							"parameter_shortname" : "live.toggle",
 							"parameter_longname" : "live.toggle",
+							"parameter_shortname" : "live.toggle",
+							"parameter_type" : 2,
+							"parameter_mmax" : 1.0,
 							"parameter_enum" : [ "off", "on" ]
 						}
 
@@ -126,7 +152,8 @@
 					"numinlets" : 2,
 					"numoutlets" : 1,
 					"outlettype" : [ "int" ],
-					"patching_rect" : [ 162.0, 53.0, 32.5, 20.0 ],
+					"patching_rect" : [ 162.0, 53.0, 32.5, 22.0 ],
+					"style" : "",
 					"text" : "- 1"
 				}
 
@@ -139,7 +166,8 @@
 					"numinlets" : 0,
 					"numoutlets" : 1,
 					"outlettype" : [ "" ],
-					"patching_rect" : [ 162.0, 20.0, 25.0, 25.0 ]
+					"patching_rect" : [ 162.0, 20.0, 25.0, 25.0 ],
+					"style" : ""
 				}
 
 			}
@@ -150,7 +178,8 @@
 					"maxclass" : "outlet",
 					"numinlets" : 1,
 					"numoutlets" : 0,
-					"patching_rect" : [ 113.25, 283.0, 25.0, 25.0 ]
+					"patching_rect" : [ 113.25, 283.0, 25.0, 25.0 ],
+					"style" : ""
 				}
 
 			}
@@ -161,7 +190,8 @@
 					"maxclass" : "outlet",
 					"numinlets" : 1,
 					"numoutlets" : 0,
-					"patching_rect" : [ 493.0, 283.0, 25.0, 25.0 ]
+					"patching_rect" : [ 493.0, 283.0, 25.0, 25.0 ],
+					"style" : ""
 				}
 
 			}
@@ -172,7 +202,8 @@
 					"maxclass" : "outlet",
 					"numinlets" : 1,
 					"numoutlets" : 0,
-					"patching_rect" : [ 16.5, 283.0, 25.0, 25.0 ]
+					"patching_rect" : [ 16.5, 283.0, 25.0, 25.0 ],
+					"style" : ""
 				}
 
 			}
@@ -185,16 +216,16 @@
 					"numinlets" : 2,
 					"numoutlets" : 1,
 					"outlettype" : [ "int" ],
-					"patching_rect" : [ 113.25, 255.0, 32.5, 20.0 ],
+					"patching_rect" : [ 113.25, 255.0, 32.5, 22.0 ],
+					"style" : "",
 					"text" : "+ 1"
 				}
 
 			}
 , 			{
 				"box" : 				{
-					"fontname" : "Arial",
+					"allowdrag" : 0,
 					"fontsize" : 9.0,
-					"framecolor" : [ 0.835415, 0.835415, 0.835415, 1.0 ],
 					"id" : "obj-3",
 					"items" : "<empty>",
 					"maxclass" : "umenu",
@@ -202,10 +233,10 @@
 					"numoutlets" : 3,
 					"outlettype" : [ "int", "", "" ],
 					"parameter_enable" : 0,
-					"patching_rect" : [ 113.25, 222.0, 146.0, 17.0 ],
+					"patching_rect" : [ 113.25, 222.0, 146.0, 19.0 ],
 					"presentation" : 1,
-					"presentation_rect" : [ 88.0, 73.0, 113.0, 17.0 ],
-					"rounded" : 3
+					"presentation_rect" : [ 88.0, 73.0, 113.0, 19.0 ],
+					"style" : ""
 				}
 
 			}
@@ -218,7 +249,8 @@
 					"numinlets" : 1,
 					"numoutlets" : 1,
 					"outlettype" : [ "" ],
-					"patching_rect" : [ 493.0, 76.5, 62.0, 18.0 ],
+					"patching_rect" : [ 493.0, 76.5, 62.0, 20.0 ],
+					"style" : "",
 					"text" : "loadmess 1"
 				}
 
@@ -234,6 +266,7 @@
 					"patching_rect" : [ 513.0, 112.0, 62.0, 16.0 ],
 					"presentation" : 1,
 					"presentation_rect" : [ 108.0, 28.5, 62.0, 16.0 ],
+					"style" : "",
 					"text" : "Loop"
 				}
 
@@ -249,6 +282,7 @@
 					"patching_rect" : [ 40.0, 112.0, 83.0, 16.0 ],
 					"presentation" : 1,
 					"presentation_rect" : [ 108.0, 6.0, 60.0, 16.0 ],
+					"style" : "",
 					"text" : "Start | Stop"
 				}
 
@@ -261,13 +295,15 @@
 					"numinlets" : 0,
 					"numoutlets" : 1,
 					"outlettype" : [ "" ],
-					"patching_rect" : [ 19.0, 20.0, 25.0, 25.0 ]
+					"patching_rect" : [ 19.0, 20.0, 25.0, 25.0 ],
+					"style" : ""
 				}
 
 			}
 , 			{
 				"box" : 				{
 					"border" : 0.0,
+					"folderslash" : 0,
 					"id" : "obj-35",
 					"maxclass" : "dropfile",
 					"numinlets" : 1,
@@ -282,18 +318,34 @@
 			}
 , 			{
 				"box" : 				{
+					"angle" : 0.0,
 					"bgcolor" : [ 0.913725, 0.913725, 0.913725, 0.0 ],
 					"border" : 1,
 					"bordercolor" : [ 0.819608, 0.819608, 0.819608, 1.0 ],
 					"id" : "obj-52",
-					"ignoreclick" : 1,
 					"maxclass" : "panel",
+					"mode" : 0,
 					"numinlets" : 1,
 					"numoutlets" : 0,
 					"patching_rect" : [ 307.25, 76.5, 132.0, 35.0 ],
 					"presentation" : 1,
 					"presentation_rect" : [ 6.0, 6.0, 80.0, 83.0 ],
-					"rounded" : 0
+					"proportion" : 0.39,
+					"rounded" : 0,
+					"style" : ""
+				}
+
+			}
+, 			{
+				"box" : 				{
+					"attr" : "folderslash",
+					"id" : "obj-4",
+					"maxclass" : "attrui",
+					"numinlets" : 1,
+					"numoutlets" : 1,
+					"outlettype" : [ "" ],
+					"patching_rect" : [ 310.0, 46.5, 150.0, 22.0 ],
+					"style" : ""
 				}
 
 			}
@@ -340,6 +392,15 @@
 					"disabled" : 0,
 					"hidden" : 0,
 					"source" : [ "obj-35", 0 ]
+				}
+
+			}
+, 			{
+				"patchline" : 				{
+					"destination" : [ "obj-35", 0 ],
+					"disabled" : 0,
+					"hidden" : 0,
+					"source" : [ "obj-4", 0 ]
 				}
 
 			}
@@ -399,13 +460,23 @@
 				}
 
 			}
+, 			{
+				"patchline" : 				{
+					"destination" : [ "obj-8", 1 ],
+					"disabled" : 0,
+					"hidden" : 0,
+					"source" : [ "obj-92", 0 ]
+				}
+
+			}
  ],
 		"parameters" : 		{
-			"obj-92" : [ "live.toggle[2]", "live.toggle", 0 ],
-			"obj-88" : [ "live.toggle", "live.toggle", 0 ]
+			"obj-88" : [ "live.toggle", "live.toggle", 0 ],
+			"obj-92" : [ "live.toggle[2]", "live.toggle", 0 ]
 		}
 ,
-		"dependency_cache" : [  ]
+		"dependency_cache" : [  ],
+		"embedsnapshot" : 0
 	}
 
 }
